:truck: Move `EvmVersionEnum` into core submodule
