+++ woke/woke/core/enums.py
@@ -1,0 +1,13 @@
+from strenum import StrEnum
+
+
+class EvmVersionEnum(StrEnum):
+    HOMESTEAD = "homestead"
+    TANGERINE_WHISTLE = "tangerineWhistle"
+    SPURIOUS_DRAGON = "spuriousDragon"
+    BYZANTIUM = "byzantium"
+    CONSTANTINOPLE = "constantinople"
+    PETERSBURG = "petersburg"
+    ISTANBUL = "istanbul"
+    BERLIN = "berlin"
+    LONDON = "london"
